refactor spark assets to asset package
