+++ great_expectations/datasource/fluent/data_asset/path/directory_asset.py
@@ -10,7 +10,7 @@ from great_expectations.core.batch import LegacyBatchDefinition
 from great_expectations.core.partitioners import Partitioner
 from great_expectations.datasource.fluent.constants import _DATA_CONNECTOR_NAME
 from great_expectations.datasource.fluent.data_asset.data_connector import FILE_PATH_BATCH_SPEC_KEY
-from great_expectations.datasource.fluent.data_asset.path_asset.file_path_data_asset import (
+from great_expectations.datasource.fluent.data_asset.path.file_path_data_asset import (
     _FilePathDataAsset,
 )
 from great_expectations.datasource.fluent.interfaces import DatasourceT
