+++ great_expectations/datasource/fluent/data_asset/path/regex_asset.py
@@ -14,7 +14,7 @@ from great_expectations.core.partitioners import (
 )
 from great_expectations.datasource.fluent.data_asset.data_connector import FILE_PATH_BATCH_SPEC_KEY
 from great_expectations.datasource.fluent.data_asset.data_connector.regex_parser import RegExParser
-from great_expectations.datasource.fluent.data_asset.path_asset.file_path_data_asset import (
+from great_expectations.datasource.fluent.data_asset.path.file_path_data_asset import (
     _FilePathDataAsset,
 )
 from great_expectations.datasource.fluent.interfaces import DatasourceT
