+++ great_expectations/datasource/fluent/spark_azure_blob_storage_datasource.py
@@ -26,7 +26,7 @@ _MISSING: Final = object()
 
 if TYPE_CHECKING:
     from great_expectations.compatibility.azure import BlobServiceClient
-    from great_expectations.datasource.fluent.spark_file_path_datasource import (
+    from great_expectations.datasource.fluent.data_asset.path.spark.spark_asset import (
         _SPARK_FILE_PATH_ASSET_TYPES_UNION,
     )
