+++ great_expectations/datasource/fluent/spark_filesystem_datasource.py
@@ -13,7 +13,7 @@ from great_expectations.datasource.fluent.data_asset.data_connector import (
 from great_expectations.datasource.fluent.interfaces import TestConnectionError
 
 if TYPE_CHECKING:
-    from great_expectations.datasource.fluent.spark_file_path_datasource import (
+    from great_expectations.datasource.fluent.data_asset.path.spark.spark_asset import (
         _SPARK_FILE_PATH_ASSET_TYPES_UNION,
     )
