+++ great_expectations/datasource/fluent/spark_google_cloud_storage_datasource.py
@@ -20,10 +20,9 @@ from great_expectations.datasource.fluent.spark_datasource import SparkDatasourc
 
 if TYPE_CHECKING:
     from great_expectations.compatibility.google import Client
-    from great_expectations.datasource.fluent.spark_file_path_datasource import (
+    from great_expectations.datasource.fluent.data_asset.path.spark.spark_asset import (
         _SPARK_FILE_PATH_ASSET_TYPES_UNION,
     )
-
 
 logger = logging.getLogger(__name__)
