+++ tests/datasource/fluent/data_asset/test_file_path_asset.py
@@ -14,6 +14,38 @@ from great_expectations.core.partitioners import (
 )
 from great_expectations.datasource.fluent import Datasource
 from great_expectations.datasource.fluent.data_asset.data_connector import FilePathDataConnector
+from great_expectations.datasource.fluent.data_asset.path.spark.csv_asset import (
+    CSVAsset as SparkCSVAsset,
+)
+from great_expectations.datasource.fluent.data_asset.path.spark.csv_asset import (
+    DirectoryCSVAsset,
+)
+from great_expectations.datasource.fluent.data_asset.path.spark.delta_asset import (
+    DeltaAsset,
+    DirectoryDeltaAsset,
+)
+from great_expectations.datasource.fluent.data_asset.path.spark.json_asset import (
+    DirectoryJSONAsset,
+)
+from great_expectations.datasource.fluent.data_asset.path.spark.json_asset import (
+    JSONAsset as SparkJSONAsset,
+)
+from great_expectations.datasource.fluent.data_asset.path.spark.orc_asset import (
+    DirectoryORCAsset,
+)
+from great_expectations.datasource.fluent.data_asset.path.spark.orc_asset import (
+    ORCAsset as SparkORCAsset,
+)
+from great_expectations.datasource.fluent.data_asset.path.spark.parquet_asset import (
+    DirectoryParquetAsset,
+)
+from great_expectations.datasource.fluent.data_asset.path.spark.parquet_asset import (
+    ParquetAsset as SparkParquetAsset,
+)
+from great_expectations.datasource.fluent.data_asset.path.spark.text_asset import (
+    DirectoryTextAsset,
+    TextAsset,
+)
 from great_expectations.datasource.fluent.file_path_asset_base import (
     AmbiguousPathError,
     PathNotFoundError,
@@ -41,28 +73,6 @@ from great_expectations.datasource.fluent.pandas_file_path_datasource import (
 )
 from great_expectations.datasource.fluent.pandas_filesystem_datasource import (
     PandasFilesystemDatasource,
-)
-from great_expectations.datasource.fluent.spark_file_path_datasource import (
-    CSVAsset as SparkCSVAsset,
-)
-from great_expectations.datasource.fluent.spark_file_path_datasource import (
-    DeltaAsset,
-    DirectoryCSVAsset,
-    DirectoryDeltaAsset,
-    DirectoryJSONAsset,
-    DirectoryORCAsset,
-    DirectoryParquetAsset,
-    DirectoryTextAsset,
-    TextAsset,
-)
-from great_expectations.datasource.fluent.spark_file_path_datasource import (
-    JSONAsset as SparkJSONAsset,
-)
-from great_expectations.datasource.fluent.spark_file_path_datasource import (
-    ORCAsset as SparkORCAsset,
-)
-from great_expectations.datasource.fluent.spark_file_path_datasource import (
-    ParquetAsset as SparkParquetAsset,
 )
