+++ tests/datasource/fluent/test_spark_dbfs_datasource.py
@@ -15,13 +15,11 @@ from great_expectations.datasource.fluent import SparkDBFSDatasource
 from great_expectations.datasource.fluent.data_asset.data_connector import (
     DBFSDataConnector,
 )
+from great_expectations.datasource.fluent.data_asset.path.spark.csv_asset import CSVAsset
 from great_expectations.datasource.fluent.file_path_data_asset import (
     _FilePathDataAsset,
 )
 from great_expectations.datasource.fluent.interfaces import TestConnectionError
-from great_expectations.datasource.fluent.spark_file_path_datasource import (
-    CSVAsset,
-)
 from tests.test_utils import create_files_in_directory
 
 if TYPE_CHECKING:
