+++ tests/datasource/fluent/test_spark_filesystem_datasource.py
@@ -22,23 +22,35 @@ from great_expectations.core.partitioners import (
 from great_expectations.datasource.fluent.data_asset.data_connector import (
     FilesystemDataConnector,
 )
+from great_expectations.datasource.fluent.data_asset.path.spark.csv_asset import (
+    CSVAsset,
+    DirectoryCSVAsset,
+)
+from great_expectations.datasource.fluent.data_asset.path.spark.delta_asset import (
+    DeltaAsset,
+    DirectoryDeltaAsset,
+)
+from great_expectations.datasource.fluent.data_asset.path.spark.json_asset import (
+    DirectoryJSONAsset,
+    JSONAsset,
+)
+from great_expectations.datasource.fluent.data_asset.path.spark.orc_asset import (
+    DirectoryORCAsset,
+    ORCAsset,
+)
+from great_expectations.datasource.fluent.data_asset.path.spark.parquet_asset import (
+    DirectoryParquetAsset,
+    ParquetAsset,
+)
+from great_expectations.datasource.fluent.data_asset.path.spark.text_asset import (
+    DirectoryTextAsset,
+    TextAsset,
+)
 from great_expectations.datasource.fluent.file_path_data_asset import _FilePathDataAsset
 from great_expectations.datasource.fluent.interfaces import (
     TestConnectionError,
 )
 from great_expectations.datasource.fluent.spark_file_path_datasource import (
-    CSVAsset,
-    DeltaAsset,
-    DirectoryCSVAsset,
-    DirectoryDeltaAsset,
-    DirectoryJSONAsset,
-    DirectoryORCAsset,
-    DirectoryParquetAsset,
-    DirectoryTextAsset,
-    JSONAsset,
-    ORCAsset,
-    ParquetAsset,
-    TextAsset,
     _SparkFilePathDatasource,
 )
 from great_expectations.datasource.fluent.spark_filesystem_datasource import (
