+++ tests/datasource/fluent/test_spark_s3_datasource.py
@@ -13,11 +13,11 @@ from great_expectations.datasource.fluent import SparkS3Datasource
 from great_expectations.datasource.fluent.data_asset.data_connector import (
     S3DataConnector,
 )
+from great_expectations.datasource.fluent.data_asset.path.spark.csv_asset import CSVAsset
 from great_expectations.datasource.fluent.file_path_data_asset import (
     _FilePathDataAsset,
 )
 from great_expectations.datasource.fluent.interfaces import TestConnectionError
-from great_expectations.datasource.fluent.spark_file_path_datasource import CSVAsset
 
 if TYPE_CHECKING:
     from botocore.client import BaseClient
